requirements.txt
numpy
Pillow
gradio
scikit-learn
